More extractions for COR-130
