+++ arangod/Cache/CMakeLists.txt
@@ -3,6 +3,7 @@ add_library(arango_cache STATIC
   Cache.cpp
   CacheManagerFeature.cpp
   CacheManagerFeatureThreads.cpp
+  CacheFeatureOptionsProvider.cpp
   CacheOptionsFeature.cpp
   CachedValue.cpp
   Finding.cpp
