+++ arangod/Cache/CacheFeatureOptionsProvider.h
@@ -1,7 +1,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 /// DISCLAIMER
 ///
-/// Copyright 2014-2024 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2014-2026 ArangoDB GmbH, Cologne, Germany
 /// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
 ///
 /// Licensed under the Business Source License 1.1 (the "License");
@@ -18,32 +18,21 @@
 ///
 /// Copyright holder is ArangoDB GmbH, Cologne, Germany
 ///
-/// @author Dan Larkin-York
 ////////////////////////////////////////////////////////////////////////////////
 
 #pragma once
 
-#include "ApplicationFeatures/ApplicationFeature.h"
+#include "ApplicationFeatures/OptionsProvider.h"
 #include "Cache/CacheOptionsProvider.h"
 
 namespace arangodb {
 
-class CacheOptionsFeature final
-    : public application_features::ApplicationFeature,
-      public CacheOptionsProvider {
- public:
-  static constexpr std::string_view name() { return "CacheOptions"; }
+struct CacheFeatureOptionsProvider : OptionsProvider<CacheOptions> {
+  void declareOptions(std::shared_ptr<options::ProgramOptions> opts,
+                      CacheOptions& options) override;
 
-  explicit CacheOptionsFeature(application_features::ApplicationServer& server);
-  ~CacheOptionsFeature() = default;
-
-  void collectOptions(std::shared_ptr<options::ProgramOptions>) override final;
-  void validateOptions(std::shared_ptr<options::ProgramOptions>) override final;
-
-  CacheOptions getOptions() const override final;
-
- private:
-  CacheOptions _options;
+  void validateOptions(std::shared_ptr<options::ProgramOptions> opts,
+                       CacheOptions& options) override;
 };
 
 }  // namespace arangodb
