+++ arangod/Cluster/MaintenanceFeature.cpp
@@ -27,6 +27,7 @@
 
 #include "Cluster/Maintenance.h"
 #include "MaintenanceFeature.h"
+#include "Cluster/MaintenanceOptionsProvider.h"
 
 #include "Metrics/CounterBuilder.h"
 #include "Metrics/GaugeBuilder.h"
@@ -193,118 +194,19 @@ MaintenanceFeature::~MaintenanceFeature() { stop(); }
 void MaintenanceFeature::collectOptions(
     std::shared_ptr<ProgramOptions> options) {
   // Initialize default values that depend on system state
-  // Corresponds to lines 167-170 in MaintenanceFeature constructor
   _options.maintenanceThreadsMax =
       (std::max)(static_cast<uint32_t>(3),  // minThreadLimit
                  static_cast<uint32_t>(NumberOfCores::getValue() / 4 + 1));
   _options.maintenanceThreadsSlowMax = _options.maintenanceThreadsMax / 2;
 
-  options->addOption(
-      "--server.maintenance-threads",
-      "The maximum number of threads available for maintenance actions.",
-      new UInt32Parameter(&_options.maintenanceThreadsMax),
-      arangodb::options::makeFlags(
-          arangodb::options::Flags::DefaultNoComponents,
-          arangodb::options::Flags::OnDBServer,
-          arangodb::options::Flags::Uncommon,
-          arangodb::options::Flags::Dynamic));
-
-  options
-      ->addOption(
-          "--server.maximal-number-sync-shard-actions",
-          "The maximum number of SynchronizeShard actions which may be queued "
-          "at any given time.",
-          new UInt64Parameter(&_options.maximalNumberOfSyncShardActionsQueued,
-                              1, 1, std::numeric_limits<uint64_t>::max()),
-          arangodb::options::makeFlags(
-              arangodb::options::Flags::DefaultNoComponents,
-              arangodb::options::Flags::OnDBServer,
-              arangodb::options::Flags::Uncommon))
-      .setIntroducedIn(31205);
-
-  options
-      ->addOption("--server.maintenance-slow-threads",
-                  "The maximum number of threads available for slow "
-                  "maintenance actions (long SynchronizeShard and long "
-                  "EnsureIndex).",
-                  new UInt32Parameter(&_options.maintenanceThreadsSlowMax),
-                  arangodb::options::makeFlags(
-                      arangodb::options::Flags::DefaultNoComponents,
-                      arangodb::options::Flags::OnDBServer,
-                      arangodb::options::Flags::Uncommon,
-                      arangodb::options::Flags::Dynamic))
-      .setIntroducedIn(30803);
-
-  options->addOption(
-      "--server.maintenance-actions-block",
-      "The minimum number of seconds finished actions block duplicates.",
-      new Int32Parameter(&_options.secondsActionsBlock),
-      arangodb::options::makeFlags(
-          arangodb::options::Flags::DefaultNoComponents,
-          arangodb::options::Flags::OnDBServer,
-          arangodb::options::Flags::Uncommon));
-
-  options->addOption(
-      "--server.maintenance-actions-linger",
-      "The minimum number of seconds finished actions remain in the deque.",
-      new Int32Parameter(&_options.secondsActionsLinger),
-      arangodb::options::makeFlags(
-          arangodb::options::Flags::DefaultNoComponents,
-          arangodb::options::Flags::OnDBServer,
-          arangodb::options::Flags::Uncommon));
-
-  options->addOption(
-      "--cluster.resign-leadership-on-shutdown",
-      "Create a resign leader ship job for this DB-Server on shutdown.",
-      new BooleanParameter(&_options.resignLeadershipOnShutdown),
-      arangodb::options::makeFlags(
-          arangodb::options::Flags::DefaultNoComponents,
-          arangodb::options::Flags::OnDBServer,
-          arangodb::options::Flags::Uncommon));
+  MaintenanceOptionsProvider provider;
+  provider.declareOptions(options, _options);
 }
 
 void MaintenanceFeature::validateOptions(
     std::shared_ptr<ProgramOptions> options) {
-  // Corresponds to lines 260-294 in MaintenanceFeature::validateOptions
-
-  // Explanation: There must always be at least 3 maintenance threads.
-  // The first one only does actions which are labelled "fast track".
-  // The next few threads do "slower" actions, but never work on very slow
-  // actions which came into being by rescheduling. If they stumble on
-  // an actions which seems to take long, they give up and reschedule
-  // with a special slow priority, such that the action is eventually
-  // executed by the slow threads. We can configure both the total number
-  // of threads as well as the number of slow threads. The number of slow
-  // threads must always be at most N-2 if N is the total number of threads.
-  // The default for the slow threads is N/2, unless the user has used
-  // an override.
-  constexpr uint32_t minThreadLimit = 3;
-  constexpr uint32_t maxThreadLimit = 64;
-
-  if (_options.maintenanceThreadsMax < minThreadLimit) {
-    LOG_TOPIC("37726", WARN, Logger::MAINTENANCE)
-        << "Need at least" << minThreadLimit << "maintenance-threads";
-    _options.maintenanceThreadsMax = minThreadLimit;
-  } else if (_options.maintenanceThreadsMax > maxThreadLimit) {
-    LOG_TOPIC("8fb0e", WARN, Logger::MAINTENANCE)
-        << "maintenance-threads limited to " << maxThreadLimit;
-    _options.maintenanceThreadsMax = maxThreadLimit;
-  }
-  if (!options->processingResult().touched("server.maintenance-slow-threads")) {
-    _options.maintenanceThreadsSlowMax = _options.maintenanceThreadsMax / 2;
-  }
-  if (_options.maintenanceThreadsSlowMax + 2 > _options.maintenanceThreadsMax) {
-    _options.maintenanceThreadsSlowMax = _options.maintenanceThreadsMax - 2;
-    LOG_TOPIC("54251", WARN, Logger::MAINTENANCE)
-        << "maintenance-slow-threads limited to "
-        << _options.maintenanceThreadsSlowMax;
-  }
-  if (_options.maintenanceThreadsSlowMax == 0) {
-    _options.maintenanceThreadsSlowMax = 1;
-    LOG_TOPIC("54252", WARN, Logger::MAINTENANCE)
-        << "maintenance-slow-threads raised to "
-        << _options.maintenanceThreadsSlowMax;
-  }
+  MaintenanceOptionsProvider provider;
+  provider.validateOptions(options, _options);
 }
 
 void MaintenanceFeature::prepare() {
